chore: move gridded files to features/gridded
