+++ frontend/src/App.jsx
@@ -9,10 +9,10 @@ import { DataDashboard } from './components/dashboards/data_management';
 import { ForecastDashboard } from './components/dashboards/forecast';
 import { Dashboard } from './components/dashboards/retrospective';
 import { NwmdDashboard } from './components/dashboards/nwmd/index';
-import GriddedDashboard from './components/dashboards/gridded/index';
+import GriddedDashboard from './features/gridded/index';
 import { DataDashboardProvider } from './context/DataDashboardContext';
 import { ForecastDashboardProvider } from './context/ForecastDashboardContext';
-import { GriddedDashboardProvider } from './context/GriddedDashboardContext';
+import { GriddedDashboardProvider } from './features/gridded/GriddedDashboardContext';
 import { NwmdDashboardProvider } from './context/NwmdDashboardContext';
 import { RetrospectiveDashboardProvider } from './context/RetrospectiveDashboardContext';
 import { useAuth } from './hooks/useAuth';
