+++ frontend/src/features/gridded/components/GriddedTimeseriesPanel.jsx
@@ -1,7 +1,7 @@
 import Plotly from 'plotly.js-dist-min';
 import { useEffect, useRef } from 'react';
-import { useGriddedDashboard } from '../../../context/GriddedDashboardContext';
-import DashboardPanel from '../../common/dashboard/DashboardPanel';
+import DashboardPanel from '../../../components/common/dashboard/DashboardPanel';
+import { useGriddedDashboard } from '../GriddedDashboardContext';
 
 const GriddedTimeseriesPanel = () => {
   const { state } = useGriddedDashboard();
