+++ frontend/src/features/gridded/utils/variableStyles.ts
